+++ wallets/ledger/src/web-usb-hid/client.ts
@@ -1,8 +1,9 @@
 import { StdSignDoc } from '@cosmjs/amino';
 import { Algo } from '@cosmjs/proto-signing';
 import { WalletClient } from '@cosmos-kit/core';
-import Cosmos from "@ledgerhq/hw-app-cosmos";
-import { getCosmosPath, ChainIdToBech32Prefix, getCosmosApp } from './utils';
+import Cosmos from '@ledgerhq/hw-app-cosmos';
+
+import { ChainIdToBech32Prefix, getCosmosApp, getCosmosPath } from './utils';
 export class LedgerClient implements WalletClient {
   client: Cosmos;
 
@@ -22,7 +23,7 @@ export class LedgerClient implements WalletClient {
       namespace: 'cosmos',
       chainId,
       address,
-      username
+      username,
     };
   }
 
@@ -30,21 +31,24 @@ export class LedgerClient implements WalletClient {
     const prefix = ChainIdToBech32Prefix[chainId];
     if (!prefix) throw new Error(`Unsupported chainId: ${chainId}`);
 
-    if (!this.client) await this.initClient()
+    if (!this.client) await this.initClient();
 
     const path = getCosmosPath(accountIndex);
     const { address, publicKey } = await this.client.getAddress(path, prefix);
     return {
       username: username ?? path,
       address,
-      algo: "secp256k1" as Algo,
+      algo: 'secp256k1' as Algo,
       pubkey: new TextEncoder().encode(publicKey),
-      isNanoLedger: true
+      isNanoLedger: true,
     };
   }
 
   async sign(signDoc: StdSignDoc, accountIndex = 0) {
-    if (!this.client) await this.initClient()
-    return await this.client.sign(getCosmosPath(accountIndex), JSON.stringify(signDoc));
+    if (!this.client) await this.initClient();
+    return await this.client.sign(
+      getCosmosPath(accountIndex),
+      JSON.stringify(signDoc)
+    );
   }
 }