+++ wallets/okxwallet-extension/src/extension/client.ts
@@ -1,5 +1,6 @@
+import { StdSignature, StdSignDoc } from '@cosmjs/amino';
 import { Algo } from '@cosmjs/proto-signing';
-import { WalletClient } from '@cosmos-kit/core';
+import { DirectSignDoc, SignOptions, WalletClient } from '@cosmos-kit/core';
 
 import { Okxwallet } from './types';
 
@@ -32,6 +33,32 @@ export class OkxwalletClient implements WalletClient {
       algo: key.algo as Algo,
       pubkey: key.pubKey,
     };
+  }
+
+  async signAmino(
+    chainId: string,
+    signer: string,
+    signDoc: StdSignDoc,
+    signOptions?: SignOptions
+  ) {
+    return await this.client.signAmino(chainId, signer, signDoc, signOptions);
+  }
+
+  async signDirect(
+    chainId: string,
+    signer: string,
+    signDoc: DirectSignDoc,
+    signOptions?: SignOptions
+  ) {
+    return await this.client.signDirect(chainId, signer, signDoc, signOptions);
+  }
+
+  async signArbitrary(
+    chainId: string,
+    signer: string,
+    data: string | Uint8Array
+  ): Promise<StdSignature> {
+    return await this.client.signArbitrary(chainId, signer, data);
   }
 
   getOfflineSigner(chainId: string) {