+++ packages/docs/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.2](https://github.com/cosmology-tech/cosmos-kit/compare/docs@2.5.1...docs@2.5.2) (2023-09-07)
+
+**Note:** Version bump only for package docs
+
+## [2.5.1](https://github.com/cosmology-tech/cosmos-kit/compare/docs@2.5.0...docs@2.5.1) (2023-09-07)
+
+**Note:** Version bump only for package docs
+
+# [2.5.0](https://github.com/cosmology-tech/cosmos-kit/compare/docs@2.4.6...docs@2.5.0) (2023-09-05)
+
+**Note:** Version bump only for package docs
+
+## [2.4.6](https://github.com/cosmology-tech/cosmos-kit/compare/docs@2.4.5...docs@2.4.6) (2023-09-02)
+
+### Bug Fixes
+
+- useChains bug when one chain has connected ([694c02c](https://github.com/cosmology-tech/cosmos-kit/commit/694c02c46c67fb70ff959134d81de6c0ee879f10))
+
 ## [2.4.5](https://github.com/cosmology-tech/cosmos-kit/compare/docs@2.4.4...docs@2.4.5) (2023-08-28)
 
 **Note:** Version bump only for package docs