+++ packages/react-lite/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.7](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react-lite@2.4.6...@cosmos-kit/react-lite@2.4.7) (2023-09-07)
+
+**Note:** Version bump only for package @cosmos-kit/react-lite
+
+## [2.4.6](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react-lite@2.4.5...@cosmos-kit/react-lite@2.4.6) (2023-09-07)
+
+**Note:** Version bump only for package @cosmos-kit/react-lite
+
+## [2.4.5](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react-lite@2.4.4...@cosmos-kit/react-lite@2.4.5) (2023-09-02)
+
+### Bug Fixes
+
+- useChains bug when one chain has connected ([694c02c](https://github.com/cosmology-tech/cosmos-kit/commit/694c02c46c67fb70ff959134d81de6c0ee879f10))
+
 ## [2.4.4](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react-lite@2.4.3...@cosmos-kit/react-lite@2.4.4) (2023-08-28)
 
 **Note:** Version bump only for package @cosmos-kit/react-lite