+++ packages/react/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.7.2](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react@2.7.1...@cosmos-kit/react@2.7.2) (2023-09-07)
+
+**Note:** Version bump only for package @cosmos-kit/react
+
+## [2.7.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react@2.7.0...@cosmos-kit/react@2.7.1) (2023-09-07)
+
+**Note:** Version bump only for package @cosmos-kit/react
+
+# [2.7.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react@2.6.5...@cosmos-kit/react@2.7.0) (2023-09-05)
+
+**Note:** Version bump only for package @cosmos-kit/react
+
+## [2.6.5](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react@2.6.4...@cosmos-kit/react@2.6.5) (2023-09-02)
+
+**Note:** Version bump only for package @cosmos-kit/react
+
 ## [2.6.4](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/react@2.6.3...@cosmos-kit/react@2.6.4) (2023-08-28)
 
 **Note:** Version bump only for package @cosmos-kit/react