+++ wallets/cosmostation-extension/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.2](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cosmostation-extension@2.4.1...@cosmos-kit/cosmostation-extension@2.4.2) (2023-09-07)
+
+**Note:** Version bump only for package @cosmos-kit/cosmostation-extension
+
+## [2.4.1](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cosmostation-extension@2.4.0...@cosmos-kit/cosmostation-extension@2.4.1) (2023-09-07)
+
+### Bug Fixes
+
+- cosmostation double sign pop up ([78a6ff7](https://github.com/cosmology-tech/cosmos-kit/commit/78a6ff7ffd5c485f526ac0d5b60c6a6604269274))
+
+# [2.4.0](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cosmostation-extension@2.3.3...@cosmos-kit/cosmostation-extension@2.4.0) (2023-09-02)
+
+### Features
+
+- Add Missing Sign Methods to Wallet Clients ([#1](https://github.com/cosmology-tech/cosmos-kit/issues/1)) ([0657840](https://github.com/cosmology-tech/cosmos-kit/commit/06578403a64a210023943031c40a9caf70d49866))
+
 ## [2.3.3](https://github.com/cosmology-tech/cosmos-kit/compare/@cosmos-kit/cosmostation-extension@2.3.2...@cosmos-kit/cosmostation-extension@2.3.3) (2023-08-24)
 
 **Note:** Version bump only for package @cosmos-kit/cosmostation-extension